Refactoring and introduce coroutine for mailing list handlers

+++ src/main/kotlin/mixit/about/AboutHandler.kt
@@ -25,7 +25,7 @@ class AboutHandler(val userService: UserService, val eventService: EventService)
 
     suspend fun findAboutView(req: ServerRequest): ServerResponse {
         val users = userService.coFindByRoles(Role.STAFF, Role.STAFF_IN_PAUSE)
-        val event = eventService.coFindByYear(AdminEventHandler.CURRENT_EVENT.toInt())
+        val event = eventService.coFindByYear(AdminEventHandler.CURRENT_EVENT)
 
         val staff = users.filter { it.role == Role.STAFF }.shuffled()
         val staffInPause = users.filter { it.role == Role.STAFF_IN_PAUSE }.shuffled()
@@ -51,7 +51,7 @@ class AboutHandler(val userService: UserService, val eventService: EventService)
             .awaitSingle()
 
     suspend fun comeToMixitView(req: ServerRequest): ServerResponse {
-        val event = eventService.coFindByYear(AdminEventHandler.CURRENT_EVENT.toInt())
+        val event = eventService.coFindByYear(AdminEventHandler.CURRENT_EVENT)
         val goldSponsors = event.filterBySponsorLevel(SponsorshipLevel.GOLD)
 
         return ok()
